Protype-S-ai (النسخة الاساسية)/pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "celery>=5.4.0",
    "elasticsearch>=8.17.2",
    "flask>=3.1.0",
    "google-generativeai>=0.8.4",
    "matplotlib>=3.10.1",
    "networkx>=3.4.2",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "serpapi>=0.1.5",
    "spacy>=3.8.4",
    "waitress>=3.0.2",
    "torch>=2.0.0",
    "torch-geometric>=2.3.0"
]